chatbridge/impl/discord/requirements.txt
discord.py
google_trans_new
pillow
imgbbpy